Clarify test and move slnf tests to a single class
